+++ src/main/kotlin/fr/insalyon/pld/agile/model/Junction.kt
@@ -6,7 +6,7 @@ import fr.insalyon.pld.agile.lib.graph.model.Measurable
  * A part of a road between two intersections
  */
 class Junction(
-    override val length: Int,
+    override val length: Long,
     val name: String
 ) : Measurable {
